perbaikan

+++ laravel/app/Http/Controllers/PenerimaanRekamController.php
@@ -358,26 +358,28 @@ class PenerimaanRekamController extends Controller {
 		$nilai = str_replace(',', '', $request->input('nilai'));
 		$pajak = 0;
 		$arr_pajak = $request->input('rincian');
-		if(count($arr_pajak)>0){
+		if(is_array($arr_pajak)){
+			if(count($arr_pajak)>0){
 			
-			$arr_keys = array_keys($arr_pajak);
-			
-			for($i=0;$i<count($arr_keys);$i++){
-				$pajak1 = str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"]);
-				$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
+				$arr_keys = array_keys($arr_pajak);
 				
-				if(isset($arr_akun[1])){
+				for($i=0;$i<count($arr_keys);$i++){
+					$pajak1 = str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"]);
+					$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
 					
-					$kddk = $arr_akun[1];
-					if($pajak1>0){
-						if($kddk=='D'){
-							$pajak += $pajak1;
+					if(isset($arr_akun[1])){
+						
+						$kddk = $arr_akun[1];
+						if($pajak1>0){
+							if($kddk=='D'){
+								$pajak += $pajak1;
+							}
+							else{
+								$pajak -= $pajak1;
+							}
 						}
-						else{
-							$pajak -= $pajak1;
-						}
+						
 					}
-					
 				}
 			}
 		}
@@ -440,47 +442,49 @@ class PenerimaanRekamController extends Controller {
 						
 						$lanjut = true;
 						$arr_pajak = $request->input('rincian');
-						if(count($arr_pajak)>0){
-							
-							$arr_keys = array_keys($arr_pajak);
-							$arr_insert = array();
-							
-							for($i=0;$i<count($arr_keys);$i++){
+						if(is_array($arr_pajak)){
+							if(count($arr_pajak)>0){
 								
-								if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+								$arr_keys = array_keys($arr_pajak);
+								$arr_insert = array();
 								
-									$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
-									$kdakun = $arr_akun[0];
+								for($i=0;$i<count($arr_keys);$i++){
+									
+									if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+									
+										$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
+										$kdakun = $arr_akun[0];
+									
+										$arr_insert[] = "select	".$id_trans." as id_trans,
+																'".$kdakun."' as kdakun,
+																".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
+														 from dual";
+														 
+									}
+									
+								}
 								
-									$arr_insert[] = "select	".$id_trans." as id_trans,
-															'".$kdakun."' as kdakun,
-															".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
-													 from dual";
-													 
+								if(count($arr_insert)>0){
+										
+									$delete = DB::delete("
+										delete from d_trans_pajak
+										where id_trans=?
+									",[
+										$id_trans
+									]);
+										
+									$insert = DB::insert("
+										insert into d_trans_pajak(id_trans,kdakun,nilai)
+										".implode(" union all ", $arr_insert)."
+									");
+									
+									if(!$insert){
+										$lanjut = false;
+									}
+									
 								}
 								
 							}
-							
-							if(count($arr_insert)>0){
-									
-								$delete = DB::delete("
-									delete from d_trans_pajak
-									where id_trans=?
-								",[
-									$id_trans
-								]);
-									
-								$insert = DB::insert("
-									insert into d_trans_pajak(id_trans,kdakun,nilai)
-									".implode(" union all ", $arr_insert)."
-								");
-								
-								if(!$insert){
-									$lanjut = false;
-								}
-								
-							}
-							
 						}
 						
 						if($lanjut){
@@ -540,47 +544,49 @@ class PenerimaanRekamController extends Controller {
 					$lanjut = true;
 					$arr_pajak = $request->input('rincian');
 					$id_trans = $request->input('inp-id');
-					if(count($arr_pajak)>0){
-						
-						$arr_keys = array_keys($arr_pajak);
-						$arr_insert = array();
-						
-						for($i=0;$i<count($arr_keys);$i++){
+					if(is_array($arr_pajak)){
+						if(count($arr_pajak)>0){
 							
-							if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+							$arr_keys = array_keys($arr_pajak);
+							$arr_insert = array();
 							
-								$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
-								$kdakun = $arr_akun[0];
+							for($i=0;$i<count($arr_keys);$i++){
+								
+								if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+								
+									$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
+									$kdakun = $arr_akun[0];
+								
+									$arr_insert[] = "select	".$id_trans." as id_trans,
+															'".$kdakun."' as kdakun,
+															".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
+													 from dual";
+													 
+								}
+								
+							}
 							
-								$arr_insert[] = "select	".$id_trans." as id_trans,
-														'".$kdakun."' as kdakun,
-														".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
-												 from dual";
-												 
+							if(count($arr_insert)>0){
+									
+								$delete = DB::delete("
+									delete from d_trans_pajak
+									where id_trans=?
+								",[
+									$id_trans
+								]);
+									
+								$insert = DB::insert("
+									insert into d_trans_pajak(id_trans,kdakun,nilai)
+									".implode(" union all ", $arr_insert)."
+								");
+								
+								if(!$insert){
+									$lanjut = false;
+								}
+								
 							}
 							
 						}
-						
-						if(count($arr_insert)>0){
-								
-							$delete = DB::delete("
-								delete from d_trans_pajak
-								where id_trans=?
-							",[
-								$id_trans
-							]);
-								
-							$insert = DB::insert("
-								insert into d_trans_pajak(id_trans,kdakun,nilai)
-								".implode(" union all ", $arr_insert)."
-							");
-							
-							if(!$insert){
-								$lanjut = false;
-							}
-							
-						}
-						
 					}
 					
 					if($lanjut){
@@ -698,12 +704,11 @@ class PenerimaanRekamController extends Controller {
 							to_char(a.tgdok,'dd-mm-yyyy') as tgpks,
 							a.id_penerima as id_pelanggan,
 							a.uraian,
-							nvl(c.nilai,0) as nilai,
-							c.kdakun,
+							nvl(a.nilai,0) as nilai,
+							a.kredit as kdakun,
 							a.id_proyek
 					from d_trans a
-					left outer join d_trans_akun c on(a.id=c.id_trans)
-					where a.id_alur=? and c.kddk='D'
+					where a.kdtran=?
 				",[
 					$parent_id
 				]);
