+++ laravel/app/Http/Controllers/PengeluaranPajakController.php
@@ -21,10 +21,10 @@ class PengeluaranPajakController extends Controller {
 							a.nodok as pks,
 							to_char(a.tgdok1,'dd-mm-yyyy') as tgjtempo,
 							a.uraian,
-							nvl(a.nilai,0) as nilai,
+							nvl(a.nilai_bersih,0) as nilai,
 							b.nmalur||'<br>'||g.nmlevel||'<br>'||c.nmstatus as status,
 							nvl(a.nilai,0)-nvl(a.nilai_bersih,0) as pajak,
-							nvl(a.nilai,0)+nvl(a.ppn,0)+nvl(a.pph21,0)+nvl(a.pph22,0)+nvl(a.pph23,0)+nvl(a.pph25,0) as total
+							nvl(a.nilai,0) as total
 					from d_trans a
 					left outer join t_alur b on(a.id_alur=b.id)
 					left outer join t_alur_status c on(a.id_alur=c.id_alur and a.status=c.status)
@@ -285,47 +285,49 @@ class PengeluaranPajakController extends Controller {
 				$lanjut = true;
 				$arr_pajak = $request->input('rincian');
 				$id_trans = $request->input('inp-id');
-				if(count($arr_pajak)>0){
-					
-					$arr_keys = array_keys($arr_pajak);
-					$arr_insert = array();
-					
-					for($i=0;$i<count($arr_keys);$i++){
+				if(is_array($arr_pajak)){
+					if(count($arr_pajak)>0){
 						
-						if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+						$arr_keys = array_keys($arr_pajak);
+						$arr_insert = array();
 						
-							$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
-							$kdakun = $arr_akun[0];
+						for($i=0;$i<count($arr_keys);$i++){
+							
+							if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+							
+								$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
+								$kdakun = $arr_akun[0];
+							
+								$arr_insert[] = "select	".$id_trans." as id_trans,
+														'".$kdakun."' as kdakun,
+														".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
+												 from dual";
+												 
+							}
+							
+						}
 						
-							$arr_insert[] = "select	".$id_trans." as id_trans,
-													'".$kdakun."' as kdakun,
-													".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
-											 from dual";
-											 
+						if(count($arr_insert)>0){
+								
+							$delete = DB::delete("
+								delete from d_trans_pajak
+								where id_trans=?
+							",[
+								$id_trans
+							]);
+								
+							$insert = DB::insert("
+								insert into d_trans_pajak(id_trans,kdakun,nilai)
+								".implode(" union all ", $arr_insert)."
+							");
+							
+							if(!$insert){
+								$lanjut = false;
+							}
+							
 						}
 						
 					}
-					
-					if(count($arr_insert)>0){
-							
-						$delete = DB::delete("
-							delete from d_trans_pajak
-							where id_trans=?
-						",[
-							$id_trans
-						]);
-							
-						$insert = DB::insert("
-							insert into d_trans_pajak(id_trans,kdakun,nilai)
-							".implode(" union all ", $arr_insert)."
-						");
-						
-						if(!$insert){
-							$lanjut = false;
-						}
-						
-					}
-					
 				}
 				
 				if($lanjut){
