+++ laravel/app/Http/Controllers/TagihanRekamController.php
@@ -313,47 +313,49 @@ class TagihanRekamController extends Controller {
 						
 						$lanjut = true;
 						$arr_pajak = $request->input('rincian');
-						if(count($arr_pajak)>0){
-							
-							$arr_keys = array_keys($arr_pajak);
-							$arr_insert = array();
-							
-							for($i=0;$i<count($arr_keys);$i++){
+						if(is_array($arr_pajak)){
+							if(count($arr_pajak)>0){
 								
-								if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+								$arr_keys = array_keys($arr_pajak);
+								$arr_insert = array();
 								
-									$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
-									$kdakun = $arr_akun[0];
+								for($i=0;$i<count($arr_keys);$i++){
+									
+									if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+									
+										$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
+										$kdakun = $arr_akun[0];
+									
+										$arr_insert[] = "select	".$id_trans." as id_trans,
+																'".$kdakun."' as kdakun,
+																".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
+														 from dual";
+														 
+									}
+									
+								}
 								
-									$arr_insert[] = "select	".$id_trans." as id_trans,
-															'".$kdakun."' as kdakun,
-															".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
-													 from dual";
-													 
+								if(count($arr_insert)>0){
+										
+									$delete = DB::delete("
+										delete from d_trans_pajak
+										where id_trans=?
+									",[
+										$id_trans
+									]);
+										
+									$insert = DB::insert("
+										insert into d_trans_pajak(id_trans,kdakun,nilai)
+										".implode(" union all ", $arr_insert)."
+									");
+									
+									if(!$insert){
+										$lanjut = false;
+									}
+									
 								}
 								
 							}
-							
-							if(count($arr_insert)>0){
-									
-								$delete = DB::delete("
-									delete from d_trans_pajak
-									where id_trans=?
-								",[
-									$id_trans
-								]);
-									
-								$insert = DB::insert("
-									insert into d_trans_pajak(id_trans,kdakun,nilai)
-									".implode(" union all ", $arr_insert)."
-								");
-								
-								if(!$insert){
-									$lanjut = false;
-								}
-								
-							}
-							
 						}
 						
 						if($lanjut){
@@ -413,47 +415,49 @@ class TagihanRekamController extends Controller {
 					$lanjut = true;
 					$arr_pajak = $request->input('rincian');
 					$id_trans = $request->input('inp-id');
-					if(count($arr_pajak)>0){
-						
-						$arr_keys = array_keys($arr_pajak);
-						$arr_insert = array();
-						
-						for($i=0;$i<count($arr_keys);$i++){
+					if(is_array($arr_pajak)){
+						if(count($arr_pajak)>0){
 							
-							if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+							$arr_keys = array_keys($arr_pajak);
+							$arr_insert = array();
 							
-								$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
-								$kdakun = $arr_akun[0];
+							for($i=0;$i<count($arr_keys);$i++){
+								
+								if($arr_pajak[$arr_keys[$i]]["'nilai'"]>0){
+								
+									$arr_akun = explode("|", $arr_pajak[$arr_keys[$i]]["'kdakun'"]);
+									$kdakun = $arr_akun[0];
+								
+									$arr_insert[] = "select	".$id_trans." as id_trans,
+															'".$kdakun."' as kdakun,
+															".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
+													 from dual";
+													 
+								}
+								
+							}
 							
-								$arr_insert[] = "select	".$id_trans." as id_trans,
-														'".$kdakun."' as kdakun,
-														".str_replace(',', '', $arr_pajak[$arr_keys[$i]]["'nilai'"])." as nilai
-												 from dual";
-												 
+							if(count($arr_insert)>0){
+									
+								$delete = DB::delete("
+									delete from d_trans_pajak
+									where id_trans=?
+								",[
+									$id_trans
+								]);
+									
+								$insert = DB::insert("
+									insert into d_trans_pajak(id_trans,kdakun,nilai)
+									".implode(" union all ", $arr_insert)."
+								");
+								
+								if(!$insert){
+									$lanjut = false;
+								}
+								
 							}
 							
 						}
-						
-						if(count($arr_insert)>0){
-								
-							$delete = DB::delete("
-								delete from d_trans_pajak
-								where id_trans=?
-							",[
-								$id_trans
-							]);
-								
-							$insert = DB::insert("
-								insert into d_trans_pajak(id_trans,kdakun,nilai)
-								".implode(" union all ", $arr_insert)."
-							");
-							
-							if(!$insert){
-								$lanjut = false;
-							}
-							
-						}
-						
 					}
 					
 					if($lanjut){
